# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "Citation": {
        "properties": {
          "excerpt": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Excerpt",
            "type": "string"
          },
          "page": {
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          }
        },
        "required": [
          "page",
          "excerpt"
        ],
        "title": "Citation",
        "type": "object"
      },
...
      "DemoDocumentDetail": {
        "properties": {
          "detected_language": {
            "title": "Detected Language",
            "type": "string"
          },
          "document_type": {
            "$ref": "#/components/schemas/DocumentType"
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/ExtractionField"
            },
            "title": "Fields",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "page_count": {
            "minimum": 1.0,
            "title": "Page Count",
            "type": "integer"
          },
          "processing": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "number"
                }
              ]
            },
            "title": "Processing",
            "type": "object"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatus"
          },
          "supported_questions": {
            "items": {
              "type": "string"
            },
            "title": "Supported Questions",
            "type": "array"
          },
          "synthetic": {
            "default": true,
            "title": "Synthetic",
            "type": "boolean"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "validation_issues": {
            "items": {
              "$ref": "#/components/schemas/ValidationIssue"
            },
            "title": "Validation Issues",
            "type": "array"
          }
        },
        "required": [
          "id",
          "name",
          "document_type",
          "detected_language",
          "status",
          "page_count",
          "updated_at",
          "fields",
          "validation_issues",
          "supported_questions",
          "processing"
        ],
        "title": "DemoDocumentDetail",
        "type": "object"
      },
...
      "DocumentResponse": {
        "properties": {
          "content_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Type"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "detected_language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detected Language"
          },
          "document_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Type"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "page_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Count"
          },
          "size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Bytes"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "filename",
          "status",
          "size_bytes",
          "page_count",
          "content_type",
          "created_at",
          "updated_at"
        ],
        "title": "DocumentResponse",
        "type": "object"
      },
      "DocumentStatus": {
        "enum": [
          "pending_upload",
          "uploaded",
          "queued",
          "validating",
          "extracting",
          "indexing",
          "ready",
          "needs_review",
          "approved",
          "failed",
          "deleted"
        ],
        "title": "DocumentStatus",
        "type": "string"
      },
      "DocumentType": {
        "enum": [
          "invoice",
          "quotation",
          "purchase_order",
          "contract"
        ],
        "title": "DocumentType",
        "type": "string"
      },
      "ExtractionField": {
        "properties": {
          "citations": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "title": "Citations",
            "type": "array"
          },
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "key",
          "label",
          "value",
          "confidence",
          "citations"
        ],
        "title": "ExtractionField",
        "type": "object"
      },
...
      "GroundedAnswer": {
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "title": "Citations",
            "type": "array"
          },
          "model_name": {
            "title": "Model Name",
            "type": "string"
          },
          "prompt_version": {
            "title": "Prompt Version",
            "type": "string"
          },
          "supported": {
            "title": "Supported",
            "type": "boolean"
          }
        },
        "required": [
          "answer",
          "citations",
          "supported",
          "model_name",
          "prompt_version"
        ],
        "title": "GroundedAnswer",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PrivateAnswerResponse": {
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "citations": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Citations",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "latency_ms": {
            "title": "Latency Ms",
            "type": "integer"
          },
          "model_name": {
            "title": "Model Name",
            "type": "string"
          },
          "prompt_version": {
            "title": "Prompt Version",
            "type": "string"
          },
          "question_id": {
            "format": "uuid",
            "title": "Question Id",
            "type": "string"
          },
          "supported": {
            "title": "Supported",
            "type": "boolean"
          }
        },
        "required": [
          "question_id",
          "answer",
          "citations",
          "supported",
          "model_name",
          "prompt_version",
          "latency_ms",
          "created_at"
        ],
        "title": "PrivateAnswerResponse",
        "type": "object"
      },
      "PrivateQuestionRequest": {
        "properties": {
          "question": {
            "maxLength": 500,
            "minLength": 3,
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "question"
        ],
        "title": "PrivateQuestionRequest",
        "type": "object"
      },
...
      "QuestionRequest": {
        "properties": {
          "question": {
            "maxLength": 500,
            "minLength": 3,
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "question"
        ],
        "title": "QuestionRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "ValidationIssue": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "related_fields": {
            "items": {
              "type": "string"
            },
            "title": "Related Fields",
            "type": "array"
          },
          "severity": {
            "title": "Severity",
            "type": "string"
          },
          "source_page": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Page"
          },
          "suggested_action": {
            "title": "Suggested Action",
            "type": "string"
          }
        },
        "required": [
          "severity",
          "code",
          "description",
          "related_fields",
          "suggested_action"
        ],
        "title": "ValidationIssue",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Authorized document workflows and a bounded, synthetic public demo.",
    "title": "GulfDocs API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/demo/documents/{document_id}": {
      "get": {
        "operationId": "get_demo_document_api_v1_demo_documents__document_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DemoDocumentDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Inspect one seeded synthetic demo document",
        "tags": [
          "public demo"
        ]
      }
    },
    "/api/v1/demo/documents/{document_id}/questions": {
      "post": {
        "operationId": "ask_demo_document_api_v1_demo_documents__document_id__questions_post",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroundedAnswer"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Answer a bounded public-demo question from precomputed evidence",
        "tags": [
          "public demo"
        ]
      }
    },
    "/api/v1/documents": {
      "get": {
        "operationId": "list_documents_api_v1_documents_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DocumentResponse"
                  },
                  "title": "Response List Documents Api V1 Documents Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Documents",
        "tags": [
          "documents"
        ]
      }
    },
    "/api/v1/documents/{document_id}": {
      "delete": {
        "operationId": "delete_document_api_v1_documents__document_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Document",
        "tags": [
          "documents"
        ]
      },
      "get": {
        "operationId": "get_document_api_v1_documents__document_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Document",
        "tags": [
          "documents"
        ]
      }
    },
...
    "/api/v1/documents/{document_id}/questions": {
      "get": {
        "operationId": "list_questions_api_v1_documents__document_id__questions_get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PrivateAnswerResponse"
                  },
                  "title": "Response List Questions Api V1 Documents  Document Id  Questions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Questions",
        "tags": [
          "documents"
        ]
      },
      "post": {
        "operationId": "ask_document_api_v1_documents__document_id__questions_post",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrivateQuestionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivateAnswerResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ask Document",
        "tags": [
          "documents"
        ]
      }
    },
...
  }
}